Ajuste no slider da homepage
