+++ src/components/Carousel/index.tsx
@@ -1,11 +1,26 @@
-import "./style.css"
+import { Swiper, SwiperSlide } from 'swiper/react';
+import { Autoplay, Pagination, Navigation } from 'swiper/modules';
+import 'swiper/css';
+import 'swiper/css/pagination';
+import 'swiper/css/navigation';
+import './style.css';
 
 const Carousel = () => {
     return(
-        <div className="carousel">
-            <h1>teste</h1>
-        </div>
-    )
-}
+        <Swiper className='swiper'
+        spaceBetween={50}
+        slidesPerView={1}
+        autoplay={{delay: 5000}}
+        pagination={{clickable: true}}
+        navigation={false}
+        modules={[Autoplay, Pagination, Navigation]}
+      >
+        <SwiperSlide>Slide 1</SwiperSlide>
+        <SwiperSlide>Slide 2</SwiperSlide>
+        <SwiperSlide>Slide 3</SwiperSlide>
+        <SwiperSlide>Slide 4</SwiperSlide>
+      </Swiper>
+    );
+  };
 
 export default Carousel;
