+++ src/components/Carousel/style.css
@@ -1,4 +1,5 @@
-.carousel{
+.swiper{
+    width: 100%;
     height: 100%;
-    background-color: blue;
+    background-color: red;
 }
