+++ src/interfaces/NoticiaAPI.ts
@@ -1,0 +1,9 @@
+interface Noticia {
+    image: string;
+    title: string;
+    content: string;
+  }
+  
+export interface NoticiasResponse {
+    noticias: Noticia[];
+  }
