+++ src/pages/HomePage/style.css
@@ -1,11 +1,13 @@
 .carousel-container{
   grid-row: span 3;
   height: 70vh;
+  width: 65vw;
 }
 
 @media (max-width: 1024px) {
   .carousel-container{
-    height: 40vh;  
+    height: 40vh;
+    width: 100%;  
     grid-column: span 3;
   }
 }
